Module Make.Multipart_upload

type t

val init : (?⁠content_type:string -> ?⁠content_encoding:(string * string) -> ?⁠acl:string -> ?⁠cache_control:string -> bucket:string -> key:string -> unit -> t result) command

Initialize multipart upload

val upload_part : (t -> part_number:int -> ?⁠expect:bool -> data:string -> unit -> unit result) command

Upload a part of the file. All parts except the last part must be at least 5Mb big. All parts must have a unique part number. The final file will be assembled from all parts ordered by part number

parameter expect:

If true, the body will not be sent until a tatus has been received from the server. This incurs a delay in transfer, but avoid sending a large body, if the request is know to fail before the body is sent.

val copy_part : (t -> part_number:int -> ?⁠range:(int * int) -> bucket:string -> key:string -> unit -> unit result) command

Specify a part as a copy of an existing object in S3.

val complete : (t -> unit -> etag result) command

Complete a multipart upload. The returned string is an opaque identifier used as etag. the etag return is _NOT_ the md5

val abort : (t -> unit -> unit result) command

Abort a multipart upload. This also discards all uploaded parts.

module Stream : sig ... end

Streaming functions