Module Multipart_upload.Stream

Streaming functions

val upload_part : (t -> part_number:int -> ?expect:bool -> data:string Io.Pipe.reader -> length:int -> chunk_size:int -> unit -> unit result) command

Streaming version of upload_part.

parameter length
is the amount of data to copy

parameter chunk_size
Is the size of chunks send to s3. The system will have 2 x chunk_size byte in flight

parameter data
the streamed data. Data will not be consumed after the result is determined. If using expect, then data may not have been consumed at all, but it is up to the caller to test if data has been consumed from the input data.

see Aws_s3.S3.Make.Multipart_upload.upload_part