Module Aws_s3__Authorization

val make_presigned_url : ?⁠scheme:[ `Http | `Https ] -> ?⁠host:string -> ?⁠port:int -> credentials:Aws_s3.Credentials.t -> date:Ptime.t -> region:Aws_s3.Region.t -> path:string -> bucket:string -> verb:[ `Get | `Put ] -> duration:int -> unit -> Uri.t

This makes a presigned url that can be used to upload or download a file from s3 without any credentials other than those embedded in the url. verb should be either the string GET for download or PUT for upload.